% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffBeta.R
\name{Diff.Beta}
\alias{Diff.Beta}
\title{Computes delta method-based confidence intervals for the difference
between pairs of standardized regression coefficients}
\usage{
Diff.Beta(
  X = NULL,
  y = NULL,
  cov.x = NULL,
  cov.xy = NULL,
  var.y = NULL,
  Nobs = NULL,
  alpha = 0.05,
  MCP = c("no", "bonf", "holm", "fdr"),
  no.diffs = NULL,
  aipe = c("no", "yes"),
  digits = 3
)
}
\arguments{
\item{X}{Raw data of predictor variables
size for a planned study}

\item{y}{Raw data of outcome variable}

\item{cov.x}{Covariance matrix among predictor variables}

\item{cov.xy}{Covariance vector between predictor variables and outcomes}

\item{var.y}{Variance of the outcome variable}

\item{Nobs}{Total sample size}

\item{alpha}{Alpha-level (\eqn{\alpha}) assumed for the planned study}

\item{MCP}{Type of multiple comparison adjustment. Enter "no" for no adjustment
or if only one pair of coefficients of interest. Enter "bonf" for Bonferroni,
"holm" for Holm, or "fdr" for the false discovery rate adjustment. The latter
two options only adjust p-value, not confidence interval.}

\item{no.diffs}{Number of pairs of coefficients to be tested. Only used with
Bonferroni method. If this value is missing, it is assumed all pairs are tested.}

\item{aipe}{Whether sequential accuracy in parameter estimation is desired for
sample size planning. Enter "yes" if aipe is desired and only interval width
will be printed. Otherwise, enter "no".}

\item{digits}{Number of digits to round confidence limits and p-values to.}
}
\value{
Delta method confidence intervals and p-values for \eqn{\beta_{i}-\beta_{j}}
Delta method covariance matrix
Delta method standard error
Confidence interval width (if \code{aipe="yes"})
}
\description{
\code{Diff.Beta} returns the lower limit and upper limit of the
  delta method based confidence interval for the difference between
  standardized regression coefficients. The estimate of the difference itself
  and a p-value indicating whether the difference is statistically significant
  are also returned. In the case of multiple differences of interest, such as
  when there are more than 2 predictor variables, multiple confidence intervals
  will be returned. In this case, users can also request intervals corrected for
  multiple comparisons using the Bonferroni approach, or p-values corrected
  using the Holm or false discovery rate methods. This approach shows better
  confidence interval coverage and Type I error rates when compared to an approach
  that relies on the standard "textbook" formula to compute the covariance matrix.
}
\details{
Researchers often compare standardized regression coefficients informally
  when comparing predictors in a regression model. However, the informal approach
  can be misleading or ineffective.

  The approach implemented in \code{Diff.Beta} uses the delta method to
  appropriately calculate the standard error for the difference between two
  standardized regression coefficients. The function assumes that predictor
  variables are random as opposed to fixed, which is typical of most predictor
  variables outside of experimental conditions.

  The calculation of a confidence interval for the difference in standardized
  regression coefficients, \eqn{\beta_{j}-\beta_{k}} requires three elements
  from the delta method covariance matrix: the variance of \eqn{\beta_{j}}, the
  variance of \eqn{\beta_{k}}, and their covariance. These elements could
  also be taken from the standard covariance matrix that duplicates the
  formula for unstandardized regression coefficients using z-scores. However,
  the standard method has been shown to be biased, producing poorly
  performing confidence intervals for single standardized regression
  coefficients (see Yuan & Chan, 2011; Jones & Waller, 2013). Intervals
  for the difference in standardized coefficients formed using the standard
  method can also behave poorly.

  In addition to the estimates of \eqn{\beta_{j}-\beta_{k}}, the limits of
  the confidence interval, and the p-value, standard output also prints
  out the full delta method covariance matrix and the standard error for
  each difference.

  Regarding function inputs, researchers can enter one of two sets of
  information. First, users can enter raw data directly. Predictors
  \code{X} are entered separately from the outcome variable \code{y}, and
  sample size is derived from the raw data. The relevant covariance
  matrices are then computed internally. Alternatively, users can enter
  covariance input and sample size. Specifically, the covariance matrix among
  predictors (\code{cov.x}), the covariances among the predictors and outcome
  (\code{cov.xy}), and the variance of the outcome (\code{var.y}) are entered,
  along with the sample size (\code{Nobs}).

  When calculating confidence intervals for multiple pairs of \eqn{\beta}
  coefficients, users have several options for familywise error adjustment.
  First, users can request unadjusted intervals using \code{MCP="no"}. Second
  users can request a Bonferroni adjustment using \code{MCP="bonf"}. When using
  Bonferroni, users can also set \code{no.diffs} equal to the number of
  differences of interest when fewer than all possible differences are desired.
  Bonferroni produces "simultaneous" intervals, in which \eqn{1-\alpha}% will
  contain the population values of all differences of interest. Appropriate
  interval formation is not possible with the Holm or FDR methods, but users
  can request these methods to calculate p-values that correct for multiple
  tests, by entering \code{MCP="holm"} and \code{MCP="fdr"}, respectively.
  When only one pair of coefficients is to be compared, users should specify
  \code{MCP="no"}.

  When researchers have flexibility in sample size planning, sequential accuracy
  in parameter estimation (AIPE) can be conducted to ensure the interval width
  is as narrow as desired. Users can set \code{aipe="yes"} for the function to
  only print the width of the interval. Otherwise, set \code{aipe="no"}.
}
\examples{
library(alr4)
d <- Rateprof
Diff.Beta(X = d[,9:11], y =d$quality, MCP = "bonf", aipe = "no", digits = 3)

}
\references{
Anderson, S. F. (2023). A Confidence Interval for the Difference Between
Standardized Regression Coefficients.

Jones, J. A., & Waller, N. G. (2013). Computing Confidence Intervals for
Standardized Regression Coefficients. Psychological Methods, 18, 453.

Yuan, K. H., & Chan, W. (2011). Biases and Standard Errors of Standardized
Regression Coefficients. Psychometrika 76, 670–690.
}
\author{
Samantha F. Anderson \email{samantha.f.anderson@asu.edu}
}
